+++ db/migrate/20150814153359_add_locked_at_to_answer_sessions.rb
@@ -1,0 +1,5 @@
+class AddLockedAtToAnswerSessions < ActiveRecord::Migration
+  def change
+    add_column :answer_sessions, :locked_at, :datetime
+  end
+end