+++ test/mailers/previews/user_mailer_preview.rb
@@ -33,6 +33,11 @@ class UserMailerPreview < ActionMailer::Preview
     UserMailer.followup_survey(answer_session)
   end
 
+  def new_surveys_available
+    user = User.second
+    UserMailer.new_surveys_available(user)
+  end
+
   def encounter_digest
     owner = User.first
     UserMailer.encounter_digest(owner, 84, { "about-me" => { name: "About Me", encounters: [ { name: "Follow Up", answer_sessions_change: 50 } ] }, "more-about-me" => { name: "More About Me", encounters: [ { name: "Seasonal Follow Up", answer_sessions_change: 12 }, { name: "Second Month", answer_sessions_change: 22 } ] } })